+++ internal/api/router.go
@@ -30,12 +30,8 @@ func NewChatRoute(router gin.IRouter, discov discoveryregistry.SvcDiscoveryRegis
 	if err != nil {
 		panic(err)
 	}
-	emoticonConn, err := discov.GetConn(context.Background(), config.Config.RpcRegisterName.OpenImChatName)
-	if err != nil {
-		panic(err)
-	}
 	mw := NewMW(adminConn)
-	chat := NewChat(chatConn, adminConn, emoticonConn)
+	chat := NewChat(chatConn, adminConn)
 	account := router.Group("/account")
 	account.POST("/code/send", chat.SendVerifyCode)                      // 发送验证码
 	account.POST("/code/verify", chat.VerifyCode)                        // 校验验证码
