--- pkg/proto/emoticon_pack/emoticon.proto
@@ -1,54 +1,0 @@
-// Copyright © 2023 OpenIM open source community. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-syntax = "proto3";
-package emoticon_pack;
-option go_package = "github.com/OpenIMSDK/chat/pkg/proto/emoticon_pack";
-
-message Emoticon {
-  int64 emoticon_id = 1;
-  string imageURL = 2;
-  string user_id = 3;
-}
-
-message AddEmoticonReq {
-  string owner_id = 1;
-  string image_data = 2;
-}
-
-message AddEmoticonResp {
-}
-
-// Request to remove an emoticon from user's storage
-message RemoveEmoticonReq {
-  string user_id = 1;
-  int64 emoticon_id = 2;
-
-}
-message RemoveEmoticonResp {
-}
-
-message GetEmoticonReq {
-  string user_id = 1;
-}
-
-message GetEmoticonResp {
-  repeated Emoticon e = 1;
-}
-
-service emoticon {
-  rpc AddEmoticon(AddEmoticonReq) returns (AddEmoticonResp);
-  rpc RemoveEmoticon(RemoveEmoticonReq) returns (RemoveEmoticonResp);
-  rpc GetEmoticon(GetEmoticonReq) returns (GetEmoticonResp);
-}
-
